Commit 38annjava-OK: Ajouté tests d'intégration Jersey.
